requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
celery==5.5.1
django-celery-beat==2.8.0
django-celery-results==2.6.0 
